Split abstract.py
